Lots of stuff

+++ MobileGame/Assets/Scripts/Core/Asteroid.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class Asteroid : MonoBehaviour
 {
-    public double pendingResources; // Ressourcen, die auf Abholung warten
+    public float pendingResources; // Ressourcen, die auf Abholung warten
 
     public void TakeDamage(float amount)
     {
