+++ MobileGame/Assets/Scripts/Data/ShipStats.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 public class ShipStats : ScriptableObject
 {
     public float speed = 5f;
-    public double capacity = 100;
+    public float capacity = 100;
     public float loadDuration = 2f; // Zeit zum Beladen am Asteroiden
 }
